+++ R/vcov3_fixest.R
@@ -4,6 +4,7 @@ vcov_CR3J.fixest <- function(
     type = "CRV3",
     return_all = FALSE,
     absorb_cluster_fixef = TRUE,
+    sparse = TRUE,
     ...
 ){
 
@@ -18,18 +19,24 @@ vcov_CR3J.fixest <- function(
   #' arXiv preprint arXiv:2205.03288 (2022).
   #'
   #' @param obj An object of type fixest
-  #' @param cluster A clustering vector
+  #' @param cluster A clustering vector. Can be a character vector of
+  #' variable names or a formula.
   #' @param absorb_cluster_fixef TRUE by default. Should the cluster fixed
   #'        effects be projected out? This increases numerical stability.
   #' @param type "CRV3" or "CRV3J" following MacKinnon, Nielsen & Webb.
   #' CRV3 by default
   #' @param return_all Logical scalar, FALSE by default. Should only
   #' the vcov be returned (FALSE) or additional results (TRUE)
+  #' @param sparse Logical scalar, TRUE by default. Should sparse matrices
+  #' be used? This is recommended for large datasets as it is considerably
+  #' faster. However, for `summclust`, `sparse = FALSE` is used so that
+  #' results match Stata. The `vcov` will be identical, but the individual
+  #' `beta_jack`s will not be.
   #' @param ... other function arguments passed to 'vcov'
   #' @method vcov_CR3J fixest
   #' @importFrom stats coef weights coefficients model.matrix
   #' @importFrom dreamerr check_arg
-  #' @importFrom MASS ginv
+  #' @importFrom Matrix sparse.model.matrix
   #' @importFrom stats expand.model.frame formula model.frame model.response na.pass pt qt reformulate
   #' @export
   #'
@@ -82,6 +89,17 @@ vcov_CR3J.fixest <- function(
       "'summclust' currently only works with estimation method 'feols'."
     )
   }
+
+  if (sparse) {
+    res <- calculate_beta_jack_sparse(obj, cluster, type, absorb_cluster_fixef, return_all)
+  } else {
+    res <- calculate_beta_jack_dense(obj, cluster, type, absorb_cluster_fixef, return_all)
+  }
+
+  res
+}
+
+calculate_beta_jack_dense <- function(obj, cluster, type, absorb_cluster_fixef, return_all) {
 
   call_env <- obj$call_env
 
@@ -151,10 +169,10 @@ vcov_CR3J.fixest <- function(
 
       add_fe <- fe[, fixef_vars, drop = FALSE]
       fml_fe <- reformulate(fixef_vars, response = NULL)
-      add_fe_dummies <- model.matrix(fml_fe, model.frame(fml_fe , data = as.data.frame(add_fe)))
+      add_fe_dummies <- Matrix::sparse.model.matrix(fml_fe, model.frame(fml_fe , data = as.data.frame(add_fe)))
       # drop the intercept
-      X <- as.matrix(collapse::add_vars(as.data.frame(X), add_fe_dummies))
-
+      #X <- Matrix::Matrix(collapse::add_vars(as.data.frame(X), add_fe_dummies))
+      X <- cbind(X, add_fe_dummies)
     }
 
   }
@@ -187,7 +205,7 @@ vcov_CR3J.fixest <- function(
     res[["k"]] <- k
     res[["cluster_df"]] <- cluster_df
   } else {
-    res <- res$vcov
+    res <- as.matrix(res$vcov)
   }
 
   class(res) <- "vcov_CR3J"
@@ -196,3 +214,145 @@ vcov_CR3J.fixest <- function(
   res
 }
 
+calculate_beta_jack_sparse <- function(obj, cluster, type, absorb_cluster_fixef, return_all) {
+
+  #' calculate leave-one out regression coefficients
+  #' @importFrom utils compareVersion
+
+  installed_version <- as.character(packageVersion("fixest"))
+
+  if(!base::requireNamespace("fixest", quietly = FALSE) || compareVersion(installed_version, "0.11.1") < 0){
+    cli::cli_abort(
+      message = "`fixest` version 0.1x or higher is required. "
+    )
+  }
+
+  X <- fixest::sparse_model_matrix(obj, type = c("rhs", "fixef"), collin.rm = TRUE)
+  y <- model.matrix(obj, type = "lhs")
+
+  N <- nrow(X)
+  # k: see below
+  w <- weights(obj)
+
+  if(!is.null(w)){
+    X <- sqrt(w) * X
+    y <- sqrt(w) * y
+  }
+
+  # get the clustering variable
+  if(!inherits(cluster, "formula")){
+    cluster <- reformulate(cluster)
+  }
+  cluster_vars <- attr(terms(cluster), "term.labels")
+
+  # Grabs data from the fixest object
+  data = fetch_data(obj, "To apply 'sparse_model_matrix', ")
+
+  # Check that cluster vars are in the original estimation dataset
+  cluster_vars_in_data <- cluster_vars %in% colnames(data)
+  if (any(!(cluster_vars_in_data))) {
+    stop(paste0(
+      "The following variables are not found in the dataset used in your `feols` call: ",
+      paste(cluster_vars[!cluster_vars_in_data], collapse = ", ")
+    ))
+  }
+
+  # Assumes that length(cluster_vars) == 1 (for now; can modify later for multi-way clustering)
+  if (length(cluster_vars) > 1) stop("Only 1 cluster variable supported right now")
+
+  cluster_vec <- data[[cluster_vars]]
+
+  # drop columns as fixest
+  if(N != length(cluster_vec)){
+    cluster_vec <- cluster_vec[unlist(obj$obs_selection)]
+  }
+
+  if (length(cluster_vec) != N) {
+    rlang::abort(
+      "The number of observations in 'cluster' and 'nobs()' do not match",
+      use_cli_format = TRUE
+    )
+  }
+
+  if (any(is.na(cluster_vec))) {
+    rlang::abort(
+      "`vcov_CR3J()` cannot handle NAs in `cluster` variables that are not
+        part of the estimated model object.",
+      use_cli_format = TRUE
+    )
+  }
+
+  unique_clusters <- as.character(unique(cluster_vec))
+
+  N_g <- table(cluster_vec)[unique_clusters]
+  G <- length(unique_clusters)
+  small_sample_correction <- (G - 1) / G
+  k <- obj$nparams
+
+  # Absorb cluster fixef effect if needed
+  cluster_fixef_outprojected <- FALSE
+
+  fixef_vars <- obj$fixef_vars
+  if (length(obj$fixef_vars) > 0) {
+
+    # if the clustering variable is a cluster fixed effect & if
+    # absorb_cluster_fixef == TRUE, then demean X and y by the
+    # cluster fixed effect
+    if (absorb_cluster_fixef && (cluster_vars %in% fixef_vars)) {
+
+      cluster_fixef_outprojected <- TRUE
+      cols_X <- colnames(X)
+      cols_to_keep <- !grepl(paste0(cluster_vars, "::"), cols_X)
+
+      cluster_fixefs <- X[, !cols_to_keep, drop = FALSE]
+      X <- X[, cols_to_keep, drop = FALSE]
+
+      # Check
+      # head(fixest::demean(as.matrix(X), nlswork[[cluster_vars]]))
+
+      # Within-transform X and y
+      X <- X - cluster_fixefs %*%
+        Matrix::solve(
+          Matrix::crossprod(cluster_fixefs),
+          Matrix::crossprod(cluster_fixefs, X)
+        )
+      y <- y - cluster_fixefs %*%
+        Matrix::solve(
+          Matrix::crossprod(cluster_fixefs),
+          Matrix::crossprod(cluster_fixefs, y)
+        )
+    }
+
+  }
+
+
+  if(cluster_fixef_outprojected){
+    k <- ncol(X)
+  } else {
+    k <- obj$nparams
+  }
+
+  res <-
+    cluster_jackknife_sparse(
+      y = y,
+      X = X,
+      cluster_vec = cluster_vec,
+      type = type
+    )
+
+  res$vcov <- as.matrix(res$vcov)
+
+  if(return_all == TRUE){
+    res[["X"]] <- X
+    res[["y"]] <- y
+    res[["N"]] <- N
+    res[["k"]] <- k
+    res[["cluster_df"]] <- data.frame(g = cluster_vec)
+  } else {
+    res <- res$vcov
+  }
+
+  class(res) <- "vcov_CR3J"
+
+  res
+}