+++ R/vcov3_lm.R
@@ -17,7 +17,9 @@ vcov_CR3J.lm <- function(
   #' arXiv preprint arXiv:2205.03288 (2022).
 
   #' @param obj An object of type lm
-  #' @param cluster A clustering vector
+  #' @param cluster A clustering vector. Can be a character vector of 
+  #' variable names or a formula.. Can be a character vector of 
+  #' variable names or a formula.
   #' @param type "CRV3" or "CRV3J" following MacKinnon, Nielsen & Webb.
   #' CRV3 by default
   #' @param return_all Logical scalar, FALSE by default. Should only
@@ -26,7 +28,6 @@ vcov_CR3J.lm <- function(
   #' @method vcov_CR3J lm
   #' @importFrom stats coef weights coefficients model.matrix
   #' @importFrom dreamerr check_arg
-  #' @importFrom MASS ginv
   #' @importFrom stats expand.model.frame formula model.frame model.response na.pass pt qt reformulate
   #'
   #' @export