+++ NEWS.md
@@ -1,4 +1,11 @@
-# New features since latest release
+# summclust 0.7
+
+* Adds option to sparsify `vcov_CR3J.fixest`. This reduces the computational cost of 
+  computing CRV3 covariance matrices when the regression's design matrix 
+  gets very sparse. It can't be used within `summclust` because the `beta_jack` do not 
+  match. However, it's useful when users just want to access `vcov_CR3J.fixest`.
+
+# summclust 0.6
 
 # summclust 0.7